Crate dusk_poseidon_merkle

Poseidon Merkle Tree

Reference implementation for the Poseidon Merkle function.

Given we have an arbitrary arity of 4 -> 1, let the following parameters be defined:

const MERKLE_ARITY: usize = 4;
const WIDTH: usize = MERKLE_ARITY + 1;
const FULL_ROUNDS: usize = 8;
const PARTIAL_ROUNDS: usize = 59;

The PoseidonMerkle structure will accept a number of inputs equal to the arity.

The input is optional for any leaf. This way, we need to alternate between a "given" and "absent" state for every leaf.

Example

use dusk_poseidon_merkle::{PoseidonMerkle, Scalar};

let mut h = PoseidonMerkle::new();
for i in 0..PoseidonMerkle::arity() {
    h.push(Scalar::from(i as u64)).unwrap();
}

let hash = h.hash();
println!("{:x?}", hash.as_bytes());

Reference

Starkad and Poseidon: New Hash Functions for Zero Knowledge Proof Systems

Structs

PoseidonMerkle
The PoseidonMerkle structure will accept a number of inputs equal to the arity.

Scalar
The Scalar struct holds an integer \(s < 2^{255} \) which represents an element of \(\mathbb Z / \ell\).

Enums

Error
Possible error states for the hashing.